🤖 Refactor: 상대경로 -\> 절대경로로 수정
