+++ frontend/src/assets/icons/whiteboard/arrow/index.ts
@@ -1,4 +1,4 @@
 // arrow icons
-export { default as ArrowIcon } from './arrowIcon.svg';
-export { default as DoubleArrowIcon } from './doubleArrowIcon.svg';
-export { default as ChevronArrowIcon } from './chevronArrowIcon.svg';
+export { default as ArrowIcon } from '@/assets/icons/whiteboard/arrow/arrowIcon.svg';
+export { default as DoubleArrowIcon } from '@/assets/icons/whiteboard/arrow/doubleArrowIcon.svg';
+export { default as ChevronArrowIcon } from '@/assets/icons/whiteboard/arrow/chevronArrowIcon.svg';
