+++ frontend/src/assets/icons/whiteboard/common/index.ts
@@ -1,2 +1,2 @@
 // common icons
-export { default as BgColorIcon } from './bgColorIcon.svg';
+export { default as BgColorIcon } from '@assets/icons/whiteboard/common/bgColorIcon.svg';
