+++ frontend/src/assets/icons/whiteboard/control/index.ts
@@ -1,7 +1,7 @@
 // control icons
-export { default as RedoIcon } from './redoIcon.svg';
-export { default as UndoIcon } from './undoIcon.svg';
-export { default as ZoomInIcon } from './zoomInIcon.svg';
-export { default as ZoomOutIcon } from './zoomOutIcon.svg';
-export { default as CloseIcon } from './closeIcon.svg';
-export { default as ShareIcon } from './shareIcon.svg';
+export { default as RedoIcon } from '@/assets/icons/whiteboard/control/redoIcon.svg';
+export { default as UndoIcon } from '@/assets/icons/whiteboard/control/undoIcon.svg';
+export { default as ZoomInIcon } from '@/assets/icons/whiteboard/control/zoomInIcon.svg';
+export { default as ZoomOutIcon } from '@/assets/icons/whiteboard/control/zoomOutIcon.svg';
+export { default as CloseIcon } from '@/assets/icons/whiteboard/control/closeIcon.svg';
+export { default as ShareIcon } from '@/assets/icons/whiteboard/control/shareIcon.svg';
