+++ frontend/src/assets/icons/whiteboard/cursor/index.ts
@@ -1,3 +1,3 @@
 // cursor icons
-export { default as CursorIcon } from './cursorIcon.svg';
-export { default as HandIcon } from './handIcon.svg';
+export { default as CursorIcon } from '@/assets/icons/whiteboard/cursor/cursorIcon.svg';
+export { default as HandIcon } from '@/assets/icons/whiteboard/cursor/handIcon.svg';
