+++ frontend/src/assets/icons/whiteboard/eraser/index.ts
@@ -1,2 +1,2 @@
 // eraser icons
-export { default as EraserIcon } from './eraserIcon.svg';
+export { default as EraserIcon } from '@/assets/icons/whiteboard/eraser/eraserIcon.svg';
