+++ frontend/src/assets/icons/whiteboard/line/index.ts
@@ -1,2 +1,2 @@
 // line icons
-export { default as LineIcon } from './lineIcon.svg';
+export { default as LineIcon } from '@/assets/icons/whiteboard/line/lineIcon.svg';
