+++ frontend/src/assets/icons/whiteboard/media/index.ts
@@ -1,4 +1,4 @@
 // media icons
-export { default as ImageIcon } from './imageIcon.svg';
-export { default as VideoIcon } from './videoIcon.svg';
-export { default as YoutubeIcon } from './youtubeIcon.svg';
+export { default as ImageIcon } from '@/assets/icons/whiteboard/media/imageIcon.svg';
+export { default as VideoIcon } from '@/assets/icons/whiteboard/media/videoIcon.svg';
+export { default as YoutubeIcon } from '@/assets/icons/whiteboard/media/youtubeIcon.svg';
