+++ frontend/src/assets/icons/whiteboard/pen/index.ts
@@ -1,2 +1,2 @@
 // pen icons
-export { default as PenIcon } from './penIcon.svg';
+export { default as PenIcon } from '@/assets/icons/whiteboard/pen/penIcon.svg';
