+++ frontend/src/assets/icons/whiteboard/shape/index.ts
@@ -1,6 +1,6 @@
 // shape icons
-export { default as CircleIcon } from './circleIcon.svg';
-export { default as TriangleIcon } from './triangleIcon.svg';
-export { default as SquareIcon } from './squareIcon.svg';
-export { default as DiamondIcon } from './diamondIcon.svg';
-export { default as PentagonIcon } from './pentagonIcon.svg';
+export { default as CircleIcon } from '@/assets/icons/whiteboard/shape/circleIcon.svg';
+export { default as TriangleIcon } from '@/assets/icons/whiteboard/shape/triangleIcon.svg';
+export { default as SquareIcon } from '@/assets/icons/whiteboard/shape/squareIcon.svg';
+export { default as DiamondIcon } from '@/assets/icons/whiteboard/shape/diamondIcon.svg';
+export { default as PentagonIcon } from '@/assets/icons/whiteboard/shape/pentagonIcon.svg';
