+++ frontend/src/assets/icons/whiteboard/text/index.ts
@@ -1,18 +1,18 @@
 // text icons
-export { default as TextBoxIcon } from './textBoxIcon.svg';
+export { default as TextBoxIcon } from '@/assets/icons/whiteboard/text/textBoxIcon.svg';
 // text align icons
-export { default as AlignCenterIcon } from './alignCenterIcon.svg';
-export { default as AlignLeftIcon } from './alignLeftIcon.svg';
-export { default as AlignRightIcon } from './alignRightIcon.svg';
+export { default as AlignCenterIcon } from '@/assets/icons/whiteboard/text/alignCenterIcon.svg';
+export { default as AlignLeftIcon } from '@/assets/icons/whiteboard/text/alignLeftIcon.svg';
+export { default as AlignRightIcon } from '@/assets/icons/whiteboard/text/alignRightIcon.svg';
 // text style icons
-export { default as BoldIcon } from './boldIcon.svg';
-export { default as ItalicIcon } from './italicIcon.svg';
+export { default as BoldIcon } from '@/assets/icons/whiteboard/text/boldIcon.svg';
+export { default as ItalicIcon } from '@/assets/icons/whiteboard/text/italicIcon.svg';
 // text orderlist icons
-export { default as OrderedListIcon } from './orderedListIcon.svg';
-export { default as UnorderListIcon } from './unorderedListIcon.svg';
+export { default as OrderedListIcon } from '@/assets/icons/whiteboard/text/orderedListIcon.svg';
+export { default as UnorderListIcon } from '@/assets/icons/whiteboard/text/unorderedListIcon.svg';
 // text decorate line icons
-export { default as StrikeThroughIcon } from './strikeThroughIcon.svg';
-export { default as TextColorIcon } from './textColorIcon.svg';
-export { default as UnderlineIcon } from './underlineIcon.svg';
+export { default as StrikeThroughIcon } from '@/assets/icons/whiteboard/text/strikeThroughIcon.svg';
+export { default as TextColorIcon } from '@/assets/icons/whiteboard/text/textColorIcon.svg';
+export { default as UnderlineIcon } from '@/assets/icons/whiteboard/text/underlineIcon.svg';
 // text link icon
-export { default as LinkIcon } from './linkIcon.svg';
+export { default as LinkIcon } from '@/assets/icons/whiteboard/text/linkIcon.svg';
